Removed OBJ files
Those only get in the way when building
Animator. The correct way would probably
be to put everything in a new directory
and leave the Pristine directory pristine,
but I only commit things on this branch
for the purpose of demonstrating what
changes are necessary to build Animator.
The next commit will have the interesting
changes.